ТЗ12 первая версия

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -3,60 +3,74 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.MPA;
-import ru.yandex.practicum.filmorate.storage.DbFilmStorage;
-import ru.yandex.practicum.filmorate.storage.DbGenreStorage;
-import ru.yandex.practicum.filmorate.storage.DbLikesStorage;
-import ru.yandex.practicum.filmorate.storage.DbMpaRatingStorage;
+import ru.yandex.practicum.filmorate.storage.film.DbFilmStorage;
+import ru.yandex.practicum.filmorate.storage.genre_mpa.DbGenreStorage;
+import ru.yandex.practicum.filmorate.storage.genre_mpa.DbMpaStorage;
+import ru.yandex.practicum.filmorate.storage.likes.DbLikesStorage;
+import ru.yandex.practicum.filmorate.storage.user.DbUserStorage;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class FilmService {
-    private final DbFilmStorage dbFilmStorage;
+    private final DbFilmStorage filmStorage;
     private final DbLikesStorage likesStorage;
     private final DbGenreStorage genreStorage;
-    private final DbMpaRatingStorage mpaRatingStorage;
+    private final DbMpaStorage mpaStorage;
+    private final DbUserStorage userStorage;
 
     public Collection<Film> findAll() {
-        return  dbFilmStorage.findAll();
+        return filmStorage.findAll();
     }
 
     public Film create(Film film) {
-     /* if (mpaRatingStorage.getById(film.getMpa().getId()) == null) {
-            throw new ValidationException("Указанный MPA  не найден");
+        if (mpaStorage.getById(film.getMpa().getId()) == null) {
+            throw new ValidationException("Рейтинг MPA с id: " + mpaStorage.getById(film.getMpa().getId()) + "  не существует");
         }
-       /* if (film.getGenres() != null) {
-            dbFilmStorage.createValidate(film.getGenres());
-        }*/
-        return dbFilmStorage.create(film);
+        if (film.getGenres() != null) {
+            genreStorage.genreValidate(film.getGenres());
+        }
+        return filmStorage.create(film);
     }
 
     public Film update(Film film) {
-        return dbFilmStorage.update(film);
+        return filmStorage.update(film);
     }
 
-    public void addLike(Long filmId, Long userId) {
-        likesStorage.addLike(filmId, userId);
+    public Film getFilm(Long id) {
+        return filmStorage.getFilm(id);
+    }
+
+    public void addLike(Long userId, Long filmId) {
+        if (filmStorage.getFilm(filmId) == null) {
+            throw new NotFoundException("Фильм с id " + filmId + " не найден");
+        }
+        if (userStorage.getUser(userId) == null) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
+        }
+        likesStorage.addLike(userId, filmId);
     }
 
     public void deleteLike(Long filmId, Long userId) {
+        if (filmStorage.getFilm(filmId) == null) {
+            throw new NotFoundException("Фильм с id " + filmId + " не найден");
+        }
+        if (userStorage.getUser(userId) == null) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
+        }
         likesStorage.deleteLike(filmId, userId);
     }
 
     public List<Film> getPopularFilms(Integer count) {
-        return likesStorage.getIdPopularFilms(count)
-                .stream()
-                .map(id -> dbFilmStorage.getFilm(id))
-                .collect(Collectors.toList());
+        return filmStorage.getIdPopularFilms(count);
     }
 
     public List<Genre> getAllGenres() {
@@ -68,10 +82,13 @@ public class FilmService {
     }
 
     public List<MPA> getAllRatings() {
-        return mpaRatingStorage.getAll();
+        return mpaStorage.getAll();
     }
 
     public MPA getMpaRating(Long id) {
-        return mpaRatingStorage.getById(id);
+        if (mpaStorage.getById(id) == null) {
+            throw new NotFoundException("Рейтинг MPA с id: " + id + "  не существует");
+        }
+        return mpaStorage.getById(id);
     }
 }
